pyproject.toml
# Application/assignment repository; not a distributable package (equivalent to Poetry package-mode = false).
[project]
name = "modern-software-dev-assignments"
version = "0.1.0"
description = "Course assignments with FastAPI + SQLite starter and tooling"
authors = [
    {name = "Course", email = "course@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.23.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "ollama>=0.5.3,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "flake8>=7.3.0",
    "httpx>=0.24.0",
    "black>=24.1.0",
    "ruff>=0.4.0",
    "pre-commit>=3.6.0",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501", "B008"]
